lift cp_sequence_numbers out of framework, into alt-schema and alt
