+++ crates/sui-indexer-alt-framework/src/models/mod.rs
@@ -1,5 +1,4 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod cp_sequence_numbers;
 pub mod watermarks;
