+++ crates/sui-indexer-alt-framework/src/schema.rs
@@ -3,14 +3,6 @@
 // @generated automatically by Diesel CLI.
 
 diesel::table! {
-    cp_sequence_numbers (cp_sequence_number) {
-        cp_sequence_number -> Int8,
-        tx_lo -> Int8,
-        epoch -> Int8,
-    }
-}
-
-diesel::table! {
     watermarks (pipeline) {
         pipeline -> Text,
         epoch_hi_inclusive -> Int8,
@@ -23,4 +15,4 @@ diesel::table! {
     }
 }
 
-diesel::allow_tables_to_appear_in_same_query!(cp_sequence_numbers, watermarks,);
+diesel::allow_tables_to_appear_in_same_query!(watermarks,);
