+++ crates/sui-indexer-alt/src/handlers/ev_struct_inst.rs
@@ -3,17 +3,18 @@
 
 use std::{collections::BTreeSet, ops::Range, sync::Arc};
 
-use crate::pg_store::PgStore;
 use anyhow::{Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    models::cp_sequence_numbers::tx_interval,
     pipeline::{concurrent::Handler, Processor},
     store::Store,
     types::full_checkpoint_content::CheckpointData,
 };
 use sui_indexer_alt_schema::{events::StoredEvStructInst, schema::ev_struct_inst};
+
+use crate::handlers::cp_sequence_numbers::tx_interval;
+use crate::pg_store::PgStore;
 
 pub(crate) struct EvStructInst;
 
@@ -92,11 +93,11 @@ impl Handler for EvStructInst {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
     use crate::new_for_testing;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
         db,
-        handlers::cp_sequence_numbers::CpSequenceNumbers,
         types::{event::Event, test_checkpoint_data_builder::TestCheckpointDataBuilder},
     };
     use sui_indexer_alt_schema::MIGRATIONS;
