+++ crates/sui-indexer-alt/src/handlers/kv_epoch_starts.rs
@@ -4,12 +4,10 @@
 use std::ops::Range;
 use std::sync::Arc;
 
-use crate::pg_store::PgStore;
 use anyhow::{bail, Context, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    models::cp_sequence_numbers::epoch_interval,
     pipeline::{concurrent::Handler, Processor},
     store::Store,
     types::{
@@ -19,6 +17,9 @@ use sui_indexer_alt_framework::{
     },
 };
 use sui_indexer_alt_schema::{epochs::StoredEpochStart, schema::kv_epoch_starts};
+
+use crate::handlers::cp_sequence_numbers::epoch_interval;
+use crate::pg_store::PgStore;
 
 pub(crate) struct KvEpochStarts;
