+++ crates/sui-indexer-alt/src/handlers/tx_calls.rs
@@ -4,17 +4,18 @@
 use std::ops::Range;
 use std::sync::Arc;
 
-use crate::pg_store::PgStore;
 use anyhow::{Ok, Result};
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    models::cp_sequence_numbers::tx_interval,
     pipeline::{concurrent::Handler, Processor},
     store::Store,
     types::{full_checkpoint_content::CheckpointData, transaction::TransactionDataAPI},
 };
 use sui_indexer_alt_schema::{schema::tx_calls, transactions::StoredTxCalls};
+
+use crate::handlers::cp_sequence_numbers::tx_interval;
+use crate::pg_store::PgStore;
 
 pub(crate) struct TxCalls;
 
@@ -92,13 +93,13 @@ impl Handler for TxCalls {
 
 #[cfg(test)]
 mod tests {
+    use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
     use crate::new_for_testing;
 
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
         db,
-        handlers::cp_sequence_numbers::CpSequenceNumbers,
         types::{base_types::ObjectID, test_checkpoint_data_builder::TestCheckpointDataBuilder},
     };
     use sui_indexer_alt_schema::MIGRATIONS;
