+++ crates/sui-indexer-alt/src/handlers/tx_digests.rs
@@ -4,17 +4,18 @@
 use std::ops::Range;
 use std::sync::Arc;
 
-use crate::pg_store::PgStore;
 use anyhow::Result;
 use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
-    models::cp_sequence_numbers::tx_interval,
     pipeline::{concurrent::Handler, Processor},
     store::Store,
     types::full_checkpoint_content::CheckpointData,
 };
 use sui_indexer_alt_schema::{schema::tx_digests, transactions::StoredTxDigest};
+
+use crate::handlers::cp_sequence_numbers::tx_interval;
+use crate::pg_store::PgStore;
 
 pub(crate) struct TxDigests;
 
@@ -80,13 +81,13 @@ impl Handler for TxDigests {
 
 #[cfg(test)]
 mod tests {
+    use crate::handlers::cp_sequence_numbers::CpSequenceNumbers;
     use crate::new_for_testing;
 
     use super::*;
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
-        db, handlers::cp_sequence_numbers::CpSequenceNumbers,
-        types::test_checkpoint_data_builder::TestCheckpointDataBuilder,
+        db, types::test_checkpoint_data_builder::TestCheckpointDataBuilder,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
